Safeguard for implicit Patch0 application
Commit 02b8d8dccc0f045b0ebf81785cf6ad1056cf550e fixed %patch to only
apply the given patch/patches specified as positional arguments, given
that -P is not used at the same time.  Previously, it would also apply
Patch0 (if present) in such a case.

This implicit application of Patch0 was unintended and basically a bug
since the dawn of time, however with that finally fixed, there's a very
small (yet non-zero) risk of breaking some old specs that rely (whether
consciously or not) on this old, buggy behavior.

Silently skipping patches that were previously applied is nasty so just
fail the build if that were to happen.

Related: RHEL-6294