Safeguard against silent Patch0 omission
Commit 02b8d8dccc0f045b0ebf81785cf6ad1056cf550e fixed %patch to only
apply what's actually supplied as positional arguments or via -P.  This
commit strengthens that for Patch0 in particular such that, if declared,
it MUST be specified on a %patch line, otherwise the build will fail.

This ensures that any previous (accidental) implicit applications of
Patch0 will now cause a build failure instead of being silently skipped.

Such specs perhaps are quite rare (if existing) but better safe than
sorry.  Silent breakage, especially due to missing patches, is nasty.

Don't fail the build if all %patch lines are using -P, though, as that's
not subject to this implicit application issue.  This includes the now
deprecated %patchN syntax, too, as that's internally translated to the
-P syntax.

Basically, this reverts the above commit but replaces the original code
with an error message.

Related: RHEL-6294