Safeguard against silent Patch0 omission
Commit 02b8d8dccc0f045b0ebf81785cf6ad1056cf550e fixed %patch so that it
no longer applies Patch0 implicitly when used without the -P option.  A
side effect of this is that, if a spec happens to rely on this quirky
behavior, it will now skip Patch0 silently.

To prevent such silent regressions in existing packages, make explicit
Patch0 application mandatory.  In other words, if Patch0 is declared,
there must be a %patch0, %patch 0 or %patch -P0 line in the spec,
otherwise the build will fail.

Unless there's no ambiguous %patch N style syntax used in the spec at
all, in which case don't fail.  This exception is not mentioned in the
error message as it would just make it confusing.

Basically, this reverts the above commit but replaces the original code
with an error message.

Related: RHEL-6294