Safeguard against silent Patch0 omission
Commit 02b8d8dccc0f045b0ebf81785cf6ad1056cf550e fixed %patch so that it
no longer applies Patch0 implicitly when used without the -P option.  A
side effect of this is that, if a spec happens to rely on this quirky
behavior, it will now skip Patch0 silently.

Avoid such nasty surprises by mandating that Patch0, if declared, always
be specified in a %patch line, unless all %patch lines in the spec use
the -P option.

Basically, this reverts the above commit but replaces the original code
with an error message.

Related: RHEL-6294