Handle plugins struct pointer being NULL
This shouldn't normally happen but rpmtsPlugins() does return NULL when
the passed transaction (ts) is NULL.  Handle that in the plugin hooks
and other utility functions so that we don't try to dereference a NULL
pointer in such a case.

This is also useful for the next commit where we'll make this into a
deliberate case.