Don't run plugins on src.rpm unpacking
Source packages aren't really "installed", just unpacked, and plugins
are by design meant for transactions, so disable the tsm/fsm/psm and
scriptlet hooks for those.

Note that the init and cleanup hooks will still run, those are supposed
to only do reversible actions such as opening and closing files, and we
wouldn't avoid them if there was a binary package supplied on the rpm -i
command line anyway.

This fixes, in particular, src.rpm installations done by a regular user
(a common case) on systems with a plugin that requires root privileges,
such as the ima plugin, which would otherwise cause a spurious warning
or even failure (see RhBug:2316785).

Extending rpmtsPlugins() here would be nicer but it's public API so just
add a static wrapper.

Reuse the plugin development test, we don't have anything better at the
moment and it does the job well.