Add "make site" target for building Jekyll site
This is immediately useful for previewing documentation changes locally
before publishing but it may come in handy if we ever need to build the
production site ourselves, too.

Uses the included Gemfile to setup the Jekyll environment.  This is done
in a Podman container for maximum isolation but one can opt to use the
native Bundler installation, provided that the required gems are already
installed (see the updated README for instructions).

Since the primary website lives in a separate repo (rpm-web), allow for
reusing the build logic for that too, by specifying a custom source dir
to build from.

Fixes: #2854