Replace ENABLE_JEKYLL with PODMAN presence check
No need to have a separate ENABLE_ option just to add a make target
(site).  Instead, add a top-level PODMAN program variable and just check
for that when deciding whether to add the target.

Note that tests/CMakeLists.txt also checks for PODMAN but don't touch
that here, we're going to refactor that part in the near future anyway.