Add a man page for rpm queryformat
Most of the content is taken from the original queryformat.md page as
well as from the QUERY FORMAT section in the rpm(8) man page, first of
which is replaced by a redirect and the latter removed in this commit.

Some parts have been rewritten from scratch or updated to reflect the
actual implementation as well.

Fixes: 3618