Make the base image usable with toolbox
Add a podman shim for the test-suite and a couple of handy utilities so
that the image can be used for toolbox-based RPM development.

Also add a quick howto. This should be documented in a better place
(perhaps a separate README file?) but for now, just add it to the
Dockerfile itself.