Fix addition of legacy signatures
Commit 5630cf4bcbaae85d1b53e7a57ba6c7ed86b62e7d fixed the semantics of
--addsign to never delete signatures, just add. This worked implicitly
for v6 signatures but was only partially handled for legacy (v4 or v3)
signatures.

Legacy signatures can be added in the following ways:

1. Signing a v4 package natively (no --rpmv6)
2. Signing a v4 package with --rpmv6
3. Signing a v6 package with --rpmv4

Currently, we prevent the addition of multiple legacy signatures in the
second case only, and happily allow any number of them in the other two
cases. Since the v4 tags aren't arrays, we end up just duplicating them.

Fix this by moving the haveLegacySig() check to putSignature() so that
we catch all legacy signature additions. Make sure to keep the identical
signature warning, though, to preserve the historical behavior and make
it easier to backport this patch to the RPM 4.x series.

Note that we have to check for the correct tags in haveLegacySig() now
since v3 signatures are only added after the v4 ones in addSignature()
and we'd otherwise always skip them.

See the updated tests for the use cases changed or fixed.

Fixes: 3878