Fix v4 signature addition
Commit 5630cf4bcbaae85d1b53e7a57ba6c7ed86b62e7d fixed the semantics of
--addsign to never delete a signature and only ever add one. This works
implicitly for v6 signatures as those can be multiple in a package, but
is only partially handled for legacy (v4 and v3) signatures which can be
added in the following ways:

1. Signing a v4 package without --rpmv6 (native)
2. Signing a v4 package with --rpmv6 (compat)
3. Signing a v6 package with --rpmv4 (compat)

Currently, only the second use case is handled, by simply unsetting the
RPMSIGN_FLAG_RPMV* flags as early as in rpmSign(). The other two cases
will result in a new legacy tag being added, even if there already is
one (whether the same algorithm or not).

For this to work correctly, the haveLegacySig() check has to be done:

1. Unconditionally (i.e. not only if --rpmv6 is used)
2. For all rpmformat values (i.e. not only for < 6)

This could be done by simply moving it outside of the current rpmformat
conditional and removing the check for RPMSIGN_FLAG_RPMV6, however that
would cause a regression in the "indentical signature" warning which we
historically issued even TBD

Thus, move the check to putSignature() instead, it also makes more sense
there. Make sure to check for the correct tags (ishdr), though, so that
addSignature() still adds the v3 signature (if requested) after the v4
signature has been added.

As a nice bonus, this also fixes the case where a legacy signature would
not be added if --resign was used together with --rpmv6 (since the flags
RPMSIGN_FLAG_RPMV* were unset even if RPMSIGN_FLAG_RESIGN was set).

See the updated tests for the actual use cases changed or fixed.