Revert "Drop local implementation of xsetprogname/xgetprogname"
Turns out some platforms (such as AIX) don't implement these.

Don't compile it just yet, it needs some fixing up, which will be done
in the next commit.

Related: #4053

This reverts commit fa06b68ed7640c0e42a8a63dd9cb0bf8d4c73525.

Adapt rpmxprogname fallback to current codebase

Fix minor issues with the original code reverted in the previous commit
that prevented it from compiling cleanly (mostly around the const usage)
and actually enable it if no system implementation is available (such as
on AIX).

Note that this new source file is tiny so it's included in libmisc
instead of a new target, and librpm is linked against it.